refactor: refactor and optimize game-switching architecture - Refactor the main function into a call method to streamline execution flow. - Add a `determine_if_change_needed` computed property to optimize game-switching logic. - Update the RemoteController class to incorporate a ScreenshotManager for improved screenshot management. - Remove unused imports for better code clarity. - Introduce a new ImageModel class in the config module and remove the deprecated image_models module. - Adjust the test files to reference the new ImageModel path. - Enhance the Screenshot class methods with clearer error handling and functionality.
