+++ pyproject.toml
@@ -76,6 +76,7 @@ dependencies = [
     "pandas>=2.2.3",
     "pillow>=11.1.0",
     "playwright>=1.49.1",
+    "playwright-stealth>=1.0.6",
     "psycopg2-binary>=2.9.10",
     "pyautogui>=0.9.54",
     "pydantic>=2.10.5",
