+++ src/controller.py
@@ -7,28 +7,21 @@ import logfire
 from adbutils import AdbDevice
 from pydantic import Field, computed_field
 import pyautogui
-from playwright.sync_api import Page
-from playwright.async_api import Page as APage
+from playwright.async_api import Page
 
 from .compare import ImageComparison
-from .screenshot import GetScreen
+from .screenshot import Screenshot, ShiftPosition, ScreenshotManager
 from .types.config import ConfigModel
-from .types.database import DatabaseConfig
+
+# from .types.database import DatabaseConfig
 from .utils.get_serial import ADBDeviceManager
-from .types.output_models import Screenshot, FoundPosition, ShiftPosition
+from .types.output_models import FoundPosition
 from .notifications.discord_notify import DiscordNotify
 
 
 class RemoteController(ConfigModel):
     found_result: FoundPosition = Field(default_factory=FoundPosition)
-    database: DatabaseConfig = Field(default_factory=DatabaseConfig)
-    start_time: datetime.datetime = Field(default_factory=datetime.datetime.now)
-
-    win: int = Field(default=0, title="Win", description="Number of wins")
-    lose: int = Field(default=0, title="Lose", description="Number of loses")
-    game_switched: bool = Field(
-        default=False, title="Game Switched", description="Whether the game has been switched"
-    )
+    screenshot_manager: ScreenshotManager = Field(default_factory=ScreenshotManager)
     notified_count: int = Field(
         default=0, title="Notified", description="Whether the notification has been sent"
     )
@@ -37,24 +30,31 @@ class RemoteController(ConfigModel):
     @property
     def target_serial(self) -> str:
         adb_manager = ADBDeviceManager(host="127.0.0.1", ports=self.serials, target=self.target)
-        apps = adb_manager()
+        apps = adb_manager.get_correct_serial()
         return apps.serial
+
+    @computed_field
+    @property
+    def determine_if_change_needed(self) -> bool:
+        current_hour = datetime.datetime.now(pytz.timezone("Asia/Taipei")).hour
+        return (22 <= current_hour < 24) or (0 <= current_hour < 1)
 
     async def get_screenshot(self) -> Screenshot:
         if self.target.startswith("http"):
-            return await GetScreen.from_remote_window(url=self.target)
+            screenshot = await self.screenshot_manager.from_browser(url=self.target)
+            return screenshot
         if self.target.startswith("com"):
-            return await GetScreen.from_adb_device(url=self.target, serial=self.target_serial)
+            screenshot = await self.screenshot_manager.from_adb(
+                url=self.target, serial=self.target_serial
+            )
+            return screenshot
         # 返回 screenshot 和 shift_position，而不是 device
-        return await GetScreen.from_exist_window(window_title=self.target)
+        screenshot = await self.screenshot_manager.from_window(window_title=self.target)
+        return screenshot
 
     async def click_button(self, device_details: Screenshot) -> None:
         if self.found_result.button_x and self.found_result.button_y:
             if isinstance(device_details.device, Page):
-                device_details.device.mouse.click(
-                    x=self.found_result.button_x, y=self.found_result.button_y
-                )
-            elif isinstance(device_details.device, APage):
                 await device_details.device.mouse.click(
                     x=self.found_result.button_x, y=self.found_result.button_y
                 )
@@ -77,42 +77,39 @@ class RemoteController(ConfigModel):
             )
 
     async def switch_game(self, device_details: Screenshot) -> None:
-        total_games = self.win + self.lose
-        current_hour = datetime.datetime.now(pytz.timezone("Asia/Taipei")).hour
-        if (22 <= current_hour < 24) or (0 <= current_hour < 1):
+        if self.determine_if_change_needed:
             return
-        if isinstance(device_details.device, AdbDevice):
-            if self.game_switched is False:
-                logfire.warn("Switching Game!!")
-                # device_details.device.click(x=1600, y=630)
-                device_details.device.shell("input tap 1600 630")
-                await asyncio.sleep(5)
-                # device_details.device.click(x=1600, y=830)
-                device_details.device.shell("input tap 1600 830")
-                await asyncio.sleep(5)
-                # device_details.device.click(x=1600, y=930)
-                device_details.device.shell("input tap 1600 930")
-                await asyncio.sleep(5)
-                self.game_switched = True
+        if not isinstance(device_details.device, AdbDevice):
+            return
+        if self.notified_count == 0:
+            logfire.warn("Switching Game!!")
+            device_details.device.click(x=1600, y=630)
+            await asyncio.sleep(5)
+            device_details.device.click(x=1600, y=830)
+            await asyncio.sleep(5)
+            device_details.device.click(x=1600, y=930)
+            await asyncio.sleep(5)
 
-                logfire.info("Game has been switched.")
-                notify = DiscordNotify(
-                    title="老闆!! 我已經幫您打完王朝了 目前已切換至五對五",
-                    description=f"王朝已完成\n目前勝場: {self.win}\n目前敗場: {self.lose}\n總場數: {total_games}",
-                    target_image=device_details.screenshot,
-                )
-                await notify.send_notify()
-                self.notified_count += 1
-            if self.notified_count == 1:
-                logfire.info("Game has been completed.")
-                notify = DiscordNotify(
-                    title="老闆!! 我已經幫您打完王朝/五對五了",
-                    description=f"五對五已完成\n目前勝場: {self.win}\n目前敗場: {self.lose}\n總場數: {total_games}",
-                    target_image=device_details.screenshot,
-                )
-                await notify.send_notify()
+            # 也可以透過下面方式來 click
+            # device_details.device.shell("input tap 1600 630")
 
-    async def run_script(self) -> None:
+            logfire.info("Game has been switched.")
+            notify = DiscordNotify(
+                title="老闆!! 我已經幫您打完王朝了 目前已切換至五對五",
+                description="王朝已完成",
+                target_image=device_details.screenshot,
+            )
+            self.notified_count += 1
+        else:
+            logfire.info("Game has been completed.")
+            notify = DiscordNotify(
+                title="老闆!! 我已經幫您打完王朝/五對五了",
+                description="五對五已完成",
+                target_image=device_details.screenshot,
+            )
+        await notify.send_notify()
+
+    async def run(self) -> None:
         try:
             device_details = await self.get_screenshot()
             for config_dict in self.image_list:
@@ -147,7 +144,3 @@ class RemoteController(ConfigModel):
                 f"Error Occurred, Retrying in {_random_interval} seconds", _exc_info=True
             )
             await asyncio.sleep(_random_interval)
-
-    async def __call__(self) -> None:
-        while True:
-            await self.run_script()
