+++ tests/test_export.py
@@ -5,7 +5,7 @@ import yaml
 import pandas as pd
 import pytest
 
-from src.types.image_models import ImageModel
+from src.types.config import ImageModel
 
 
 @pytest.fixture(scope="module")
